Move inkbird coordinator logic into coordinator.py
Not a functional change, one to one relocation
